Evaluation/Base REGIS/requirements.txt
numpy==1.24.2
pandas==1.4.4
plotly==5.9.0
requests==2.28.1
scikit-learn==1.0.2
pydantic==1.10.7
python-dotenv==1.0.0
rdflib==6.3.2
pyyaml==6.0
pymongo==4.3.3
pyswarms==1.3.0
